requirements.txt
kivy==2.3.0
kivy-garden.graph==0.4.0
Pillow==10.1.0
pytesseract==0.3.10
sqlite3
plyer==2.1.0
requests==2.31.0
numpy==1.24.3
scipy==1.11.4
geopy==2.4.1
pyttsx3==2.90
speechrecognition==3.10.0
opencv-python==4.8.1.78
pygame==2.5.2